Move record batch test fixtures into test-common.h
